:wrench: 모듈 패키지명 및 namespace 변경
