--- challenge/src/main/java/com/soopeach/challenge/ChallengeFragment.kt
@@ -1,19 +1,0 @@
-package com.soopeach.challenge
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class ChallengeFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_challenge, container, false)
-    }
-
-}
