+++ community/src/main/java/com/whyranoid/community/CommunityFragment.kt
@@ -1,0 +1,18 @@
+package com.whyranoid.community
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+class CommunityFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_community, container, false)
+    }
+
+}
