+++ community/src/main/res/navigation/community_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/communityFragment"
-        android:name="com.soopeach.community.CommunityFragment"
+        android:name="com.whyranoid.community.CommunityFragment"
         android:label="fragment_community"
         tools:layout="@layout/fragment_community" />
 </navigation>
