+++ mypage/src/main/res/navigation/my_page_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.soopeach.mypage.MyPageFragment"
+        android:name="com.whyranoid.mypage.MyPageFragment"
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page" />
 </navigation>
