+++ running/src/main/java/com/whyranoid/running/RunningFragment.kt
@@ -1,0 +1,19 @@
+package com.whyranoid.running
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class RunningFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_running, container, false)
+    }
+        
+}
