+++ running/src/main/res/navigation/running_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/runningFragment"
-        android:name="com.soopeach.running.RunningFragment"
+        android:name="com.whyranoid.running.RunningFragment"
         android:label="fragment_running"
         tools:layout="@layout/fragment_running" />
 </navigation>
